Code Cleanup

+++ src/test/java/org/sorting/ArrayGenerator.java
@@ -34,11 +34,10 @@ public class ArrayGenerator
         {
             return new ArrayList<>(0);
         }
-        
         var orders = new ArrayList<int[]>((int)Math.pow(upperBound - lowerBound+1, size));
         orders.add(Arrays.stream(new int[size]).map(i -> lowerBound).toArray());
         var last = orders.getLast();
-        for (long j = 1; j < Math.pow((long)upperBound - lowerBound+1, size); j++)
+        for (long j = 1; j < Math.pow( upperBound - lowerBound+1, size); j++)
         {
             var next = last.clone();
